improved hostile entity AI pathfinding
Added a check in hostile enemy ai to verify that the the next step in the path is free. This should ensure that invalid path destinations don't get consumed when the entity is out of sight, therefore preventing enemies from moving more then one tiles when a previous step has been consumed but blocked.
This should fix #17